+++ OggConverter/src/Forms/SettingsForm.cs
@@ -29,12 +29,13 @@ namespace OggConverter
         {
             InitializeComponent();
 
-            Localize();
+            Localise();
             labVer.Text = Localisation.Get("Your Version: {0}\n" +
                 "Build: {1}", Utilities.GetVersion(true), Updates.version);
 
             tabControl.ItemSize = new Size((tabControl.Width / tabControl.TabCount) - 1, 0);
 
+            // Adding already translates elementr to youtube-dl update frequency setting
             cbYoutubeDlUpdateFrequency.Items.Add(Localisation.Get("Upon every start"));
             cbYoutubeDlUpdateFrequency.Items.Add(Localisation.Get("Daily"));
             cbYoutubeDlUpdateFrequency.Items.Add(Localisation.Get("Weekly"));
@@ -193,7 +194,6 @@ namespace OggConverter
 
         private void CbYoutubeDlUpdateFrequency_SelectedIndexChanged(object sender, EventArgs e)
         {
-            Settings.YouTubeDlUpdateFrequency = cbYoutubeDlUpdateFrequency.SelectedIndex;
         }
 
         private void BtnCheckUpdate_Click(object sender, EventArgs e)
@@ -275,7 +275,7 @@ namespace OggConverter
             }
         }
 
-        void Localize()
+        void Localise()
         {
             chkShortcut.Text = Localisation.Get("Desktop shortcut");
             chkNoSteam.Text = Localisation.Get("Start the game without Steam");
@@ -307,6 +307,12 @@ namespace OggConverter
             tabLogging.Text = Localisation.Get("Logging & Privacy");
 
             this.Text = Localisation.Get("Settings");
+        }
+
+        private void CbYoutubeDlUpdateFrequency_SelectionChangeCommitted(object sender, EventArgs e)
+        {
+            bool isLastSelected = cbYoutubeDlUpdateFrequency.SelectedIndex + 1 == cbYoutubeDlUpdateFrequency.Items.Count;
+            Settings.YouTubeDlUpdateFrequency = isLastSelected ? -1 : cbYoutubeDlUpdateFrequency.SelectedIndex;
         }
     }
 }