+++ OggConverter/src/Music/Player.cs
@@ -99,41 +99,38 @@ namespace OggConverter
 
             Stop();
 
-            int skipped = 0;
-
             try
             {
-                for (int i = 1; i <= 99; i++)
+                int lastFileNumber = 0;
+                // Getting directory info from file and saving only files that are named track*.ogg into var file
+                DirectoryInfo di = new DirectoryInfo($"{Settings.GamePath}\\{folder}");
+                var files = di.GetFileSystemInfos("track*.ogg").OrderBy(f => int.Parse(f.Name.Replace("track", "").Split('.')[0]));
+                for (int i = 0; i < files.Count(); i++)
                 {
-                    // File doesn't exist? Skipping the value
-                    if (!File.Exists($"{Settings.GamePath}\\{folder}\\track{i}.ogg"))
+                    int thisFileNumber = int.Parse(files.ToArray()[i].Name.Replace("track", "").Split('.')[0]);
+                    // If the difference between last checked file and the current one is only 1
+                    if (thisFileNumber - lastFileNumber == 1)
                     {
-                        skipped++;
+                        // Save that file as the last one and continue
+                        lastFileNumber = thisFileNumber;
+                        continue;
                     }
-                    else
-                    {
-                        // If nothing was skipped, there is no point of moving file
-                        if (skipped == 0) continue;
 
-                        // Waiting for file to be free
-                        while (!Utilities.IsFileReady($"{Settings.GamePath}\\{folder}\\track{i}.ogg")) { }
+                    // Calculates the new file number
+                    int newNumber = thisFileNumber - (thisFileNumber - lastFileNumber) + 1;
+                    // Waiting for file to be free
+                    while (!Utilities.IsFileReady($"{Settings.GamePath}\\{folder}\\track{thisFileNumber}.ogg")) { }
 
-                        // Moving the file
-                        File.Move($"{Settings.GamePath}\\{folder}\\track{i}.ogg", $"{Settings.GamePath}\\{folder}\\track{i - skipped}.ogg");
+                    // Moving the file
+                    File.Move($"{Settings.GamePath}\\{folder}\\track{thisFileNumber}.ogg",
+                        $"{Settings.GamePath}\\{folder}\\track{newNumber}.ogg");
 
-                        Logs.History(Localisation.Get("Sorting: moved 'track{0}' to 'track{1}' in {2}", i, i - skipped, folder));
-                        Form1.instance.Log(Localisation.Get("Sorting: moved 'track{0}' to 'track{1}' in {2}", i, i - skipped, folder));
+                    Logs.History(Localisation.Get("Sorting: moved 'track{0}' to 'track{1}' in {2}", thisFileNumber, newNumber, folder));
+                    Form1.instance.Log(Localisation.Get("Sorting: moved 'track{0}' to 'track{1}' in {2}", thisFileNumber, newNumber, folder));
 
-                        MetaData.AddOrEdit($"track{i - skipped}", MetaData.GetName($"track{i}"));
-                        MetaData.Remove($"track{i}");
-
-                        // Adjusting the i value by skipped
-                        if (skipped != 0)
-                            i -= skipped;
-
-                        // Reseting skipped value
-                        skipped = 0;
-                    }
+                    MetaData.AddOrEdit($"track{newNumber}", MetaData.GetName($"track{thisFileNumber}"));
+                    MetaData.Remove($"track{thisFileNumber}");
+                    lastFileNumber = newNumber;
                 }
             }
             catch (Exception ex)
@@ -157,7 +154,10 @@ namespace OggConverter
         public static void ChangeOrder(ListBox songList, string folder, bool moveUp)
         {
             int selectedIndex = songList.SelectedIndex;
-            if ((songList.SelectedIndex == -1) || (selectedIndex == 0 && moveUp) || (selectedIndex == songList.Items.Count - 1 && !moveUp)) return;
+            if ((songList.SelectedIndex == -1) || 
+                (selectedIndex == 0 && moveUp) || 
+                (selectedIndex == songList.Items.Count - 1 && !moveUp))
+                return;
 
             if (Utilities.IsToolBusy())
             {
@@ -281,14 +281,11 @@ namespace OggConverter
             try
             {
                 // Getting a new name for cloned song
-                for (int i = 1; i <= 99; i++)
-                {
-                    if (!File.Exists($"{Settings.GamePath}\\{folder}\\track{i}.ogg"))
-                    {
-                        newName = $"track{i}";
-                        break;
-                    }
-                }
+                int newNumber = 1;
+                for (int i = 1; File.Exists($"{Settings.GamePath}\\{folder}\\track{newNumber}.ogg"); i++)
+                    newNumber++;
+
+                newName = $"track{newNumber}";
 
                 string pathToFile = $"{Settings.GamePath}\\{folder}\\{fileName}"; // Path to file to be cloned with it's name